Extract shared providerLabel utility for consistent OAuth labels
Address review feedback: the OAuthCredentialRow component was using
charAt(0).toUpperCase() which produces incorrect brand names like
"Linkedin" instead of "LinkedIn". Extract a shared providerLabel()
utility with a canonical label map covering all built-in providers.
Replaces duplicate PROVIDER_LABELS maps in three files.
