+++ frontend/src/pages/settings/ConnectedAccountsSection.tsx
@@ -12,6 +12,7 @@ import { useAuth } from "@/auth/AuthContext";
 import { useOAuthConnections } from "@/hooks/useOAuthConnections";
 import { useOAuthProviders } from "@/hooks/useOAuthProviders";
 import { useDisconnectOAuth } from "@/hooks/useDisconnectOAuth";
+import { providerLabel } from "@/lib/oauthProviders";
 import { InlineConfirmButton } from "@/components/InlineConfirmButton";
 import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
@@ -22,17 +23,6 @@ import {
   CardHeader,
   CardTitle,
 } from "@/components/ui/card";
-
-const PROVIDER_LABELS: Record<string, string> = {
-  google: "Google",
-  intercom: "Intercom",
-  linear: "Linear",
-  microsoft: "Microsoft",
-};
-
-function providerLabel(id: string): string {
-  return PROVIDER_LABELS[id] ?? id.charAt(0).toUpperCase() + id.slice(1);
-}
 
 function statusBadge(status: string) {
   switch (status) {
